Add TaskResultAnalyzer to generate output. Moved most of Dartagnan.java's output generation code to TaskResultAnalyzer
